+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the **"my-racing-planner"** project will be documented in
 ## 0.6.2
 
 - Little improvement to license badge, looking better on light theme
+- Added track config to season planner table
 
 ## 0.6.1